+++ src/main/java/de/rolandsw/schedulemc/weapon/grenade/FlashGrenadeItem.java
@@ -1,0 +1,5 @@
+package de.rolandsw.schedulemc.weapon.grenade;
+
+public class FlashGrenadeItem extends GrenadeItem {
+    public FlashGrenadeItem() { super(GrenadeType.FLASH); }
+}