+++ src/main/java/de/rolandsw/schedulemc/weapon/grenade/GrenadeType.java
@@ -1,0 +1,5 @@
+package de.rolandsw.schedulemc.weapon.grenade;
+
+public enum GrenadeType {
+    FRAG, SMOKE, FLASH
+}