+++ src/main/java/de/rolandsw/schedulemc/weapon/upgrade/AutoFireUpgradeItem.java
@@ -1,0 +1,8 @@
+package de.rolandsw.schedulemc.weapon.upgrade;
+
+public class AutoFireUpgradeItem extends BaseFireModeUpgradeItem {
+    @Override
+    public FireModeUpgradeType getUpgradeType() {
+        return FireModeUpgradeType.AUTO;
+    }
+}