+++ src/main/java/de/rolandsw/schedulemc/weapon/upgrade/BurstFireUpgradeItem.java
@@ -1,0 +1,8 @@
+package de.rolandsw.schedulemc.weapon.upgrade;
+
+public class BurstFireUpgradeItem extends BaseFireModeUpgradeItem {
+    @Override
+    public FireModeUpgradeType getUpgradeType() {
+        return FireModeUpgradeType.BURST;
+    }
+}