+++ src/main/java/de/rolandsw/schedulemc/weapon/util/WeaponNBT.java
@@ -1,0 +1,11 @@
+package de.rolandsw.schedulemc.weapon.util;
+
+public class WeaponNBT {
+    public static final String ATTACHMENT_TYPE = "AttachmentType";
+    public static final String ATTACHMENT_DAMAGE = "AttachmentDamage";
+    public static final String ATTACHMENT_SPREAD = "AttachmentSpread";
+    public static final String ATTACHMENT_ACCURACY = "AttachmentAccuracy";
+    public static final String ATTACHMENT_SOUND = "AttachmentSound";
+    public static final String AUTO_FIRE_ACTIVE = "WeaponAutoFireActive";
+    public static final String AUTO_FIRE_GUN_SLOT = "WeaponAutoFireGunSlot";
+}